requirements.txt
affine==2.4.0
anyio==4.8.0
attrs==25.3.0
blosc2==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
contextily==1.6.2
contourpy==1.3.1
coverage==7.6.10
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.55.6
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.0
mercantile==1.2.1
msgpack==1.1.0
munkres==1.1.4
ndindex==1.9.2
networkx==3.4.2
numba==0.61.0
numexpr==2.10.2
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==5.24.1
pluggy==1.5.0
py-cpuinfo==9.0.0
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.1
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rasterio==1.4.3
recordclass==0.22.1
requests==2.32.3
responses==0.25.6
rtree==1.4.0
scikit-learn==1.6.1
scipy==1.15.1
score-clustering==0.1.4
seaborn==0.13.2
setuptools==75.1.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tables==3.10.2
tenacity==9.0.0
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
unicodedata2==16.0.0
urllib3==2.3.0
wheel==0.44.0
xyzservices==2025.1.0
